+++ plugins/analytics-provider-segment/src/plugin.ts
@@ -16,10 +16,8 @@
 import { createPlugin } from '@backstage/core-plugin-api';
 
 /**
- * @deprecated Importing and including this plugin in an app has no effect.
- * This will be removed in a future release.
  *
- * @public
+ * @public Importing and including this plugin in an app has no effect.
  */
 export const analyticsModuleSegment = createPlugin({
   id: 'analytics-provider-segment',