+++ plugins/kiali/src/Router.tsx
@@ -9,7 +9,7 @@ import { Button } from '@material-ui/core';
 
 import { KUBERNETES_NAMESPACE } from '@janus-idp/backstage-plugin-kiali-common';
 
-import { KialiComponent } from './components/KialiComponent';
+import { KialiComponent, KialiNoPath } from './components/KialiComponent';
 
 export const KUBERNETES_ANNOTATION = 'backstage.io/kubernetes-id';
 const KUBERNETES_LABEL_SELECTOR_QUERY_ANNOTATION =
@@ -40,6 +40,8 @@ export const Router = () => {
     return (
       <Routes>
         <Route path="/" element={<KialiComponent />} />
+        <Route path="/overview" element={<KialiComponent />} />
+        <Route path="*" element={<KialiNoPath />} />
       </Routes>
     );
   }