+++ plugins/kiali/src/components/KialiComponent/KialiComponent.tsx
@@ -22,6 +22,7 @@ import {
   KialiConfigT,
   KialiFetchError,
   KialiInfo,
+  Namespace,
   setServerConfig,
 } from '@janus-idp/backstage-plugin-kiali-common';
 
@@ -30,16 +31,51 @@ import { handleMultipleMessage } from '../../helper';
 import { Overview } from '../Overview';
 import { KialiHeader } from './Header';
 
+const getPathPage = () => {
+  const pathname = window.location.pathname.split('/').pop() || 'overview';
+  return pathname === 'kiali' ? 'overview' : pathname;
+};
+
 export const KialiComponent = () => {
   const kialiClient = useApi(kialiApiRef);
   kialiClient.setEntity(useEntity().entity);
+  const [kialiTab, setKialiTab] = React.useState<string>(getPathPage());
   const [kialiConfig, setKialiConfig] =
     React.useState<KialiConfigT>(DefaultKialiConfig);
   const [kialiStatus, setKialiStatus] = React.useState<KialiInfo>({
     status: INITIAL_STATUS_STATE,
     auth: { sessionInfo: {}, strategy: AuthStrategy.anonymous },
   });
+  const [namespacesFiltered, setNamespacesFiltered] = React.useState<string[]>(
+    [],
+  );
+  const [namespaces, setNamespaces] = React.useState<Namespace[]>([]);
   const [errors, setErrors] = React.useState<KialiFetchError[]>([]);
+
+  const fetchConfig = async () => {
+    let config = kialiConfig;
+    if (config.kialiConsole === '') {
+      await kialiClient.getConfig().then(resp => {
+        if (resp.errors.length > 0) {
+          setErrors(resp.errors);
+        }
+        config = resp.response as KialiConfigT;
+        config.server = setServerConfig(kialiConfig?.server, config.server);
+        setKialiConfig(config);
+      });
+    }
+  };
+
+  const fetchNamespaces = async () => {
+    await kialiClient.getNamespaces().then(resp => {
+      if (resp.errors.length > 0) {
+        setErrors(resp.errors);
+      }
+      setNamespaces(resp.response as Namespace[]);
+      setNamespacesFiltered((resp.response as Namespace[]).map(ns => ns.name));
+    });
+  };
+
   const [{ loading }, refresh] = useAsyncFn(
     async () => {
       await kialiClient.getInfo().then(response => {
@@ -47,20 +83,8 @@ export const KialiComponent = () => {
           setErrors(response.errors);
         } else {
           setKialiStatus(response.response as KialiInfo);
-          let config = kialiConfig;
-          if (config.kialiConsole === '') {
-            kialiClient.getConfig().then(resp => {
-              if (resp.errors.length > 0) {
-                setErrors(resp.errors);
-              }
-              config = resp.response as KialiConfigT;
-              config.server = setServerConfig(
-                kialiConfig?.server,
-                config.server,
-              );
-              setKialiConfig(config);
-            });
-          }
+          fetchConfig();
+          fetchNamespaces();
         }
       }); // Check if the config is loaded
     },
@@ -79,6 +103,9 @@ export const KialiComponent = () => {
           title="Kiali"
           kialiStatus={kialiStatus}
           config={kialiConfig}
+          namespaces={namespaces}
+          setNamespaceFilter={setNamespacesFiltered}
+          namespacesFiltered={namespacesFiltered}
         />
         {errors.length > 0 && (
           <WarningPanel
@@ -90,11 +117,18 @@ export const KialiComponent = () => {
         )}
         {kialiConfig.kialiConsole !== '' && (
           <>
-            <TabbedCard>
+            <TabbedCard
+              value={kialiTab}
+              onChange={(_, v) => setKialiTab(v as string)}
+            >
               {/* 
                 // @ts-ignore */}
-              <CardTab label="Overview">
-                <Overview kialiConfig={kialiConfig} kialiStatus={kialiStatus} />
+              <CardTab value="overview" label="Overview">
+                <Overview
+                  kialiConfig={kialiConfig}
+                  kialiStatus={kialiStatus}
+                  namespacesFiltered={namespacesFiltered}
+                />
               </CardTab>
             </TabbedCard>
           </>